Refactor list filter
